Moved IConfigurationSetting to DotNetNuke.Abstractions.Settings
