+++ DNN Platform/DotNetNuke.Abstractions/Settings/IConfigurationSetting.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information
 
-namespace DotNetNuke.Abstractions
+namespace DotNetNuke.Abstractions.Settings
 {
     /// <summary>
     /// The configuration setting.
